AI-Powered Resume Builder with Job Matching System(Sravanthi_Vaka)/requirements.txt
streamlit==1.37.1
langchain==0.2.14
langgraph==0.2.20
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
pdfplumber==0.11.4
PyPDF2==3.0.1
reportlab==4.2.2
graphviz==0.20.3
